+++ CHANGES.rst
@@ -11,6 +11,8 @@ Breaking changes
 - Alectryon now generates simpler HTML markup for hypothesis blocks (may affect third-party
   stylesheets).
 
+- ``json.Cache`` now takes arbitrary ``metadata`` instead of ``sertop_args``.
+
 Version 1.2.1
 =============